+++ CHANGELOG.md
@@ -2,7 +2,11 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+### Changed
 
+ - Populate `Content-Type` and `Content-Length` headers correctly
+   when transforming the Akka HttpResponse to the AWS SDK Response
+   
 ## [1.0.1]
 ### Changed